refactoring Split layout
git-svn-id: http://192.168.0.243:4435/svn/bee/trunk@749 aaee961a-53e7-2249-bd01-b2a8f0c4ce08

+++ src/com/butent/bee/client/logging/PanelHandler.java
@@ -132,7 +132,7 @@ public class PanelHandler extends Handler implements HasVisibility {
   private int getSize() {
     Widget parent = getPanel().getParent();
     if (parent instanceof Split) {
-      return (int) ((Split) parent).getWidgetSize(getPanel());
+      return ((Split) parent).getWidgetSize(getPanel());
     } else {
       return BeeConst.UNDEF;
     }
